parsing fix

+++ soccer/Ball.java
@@ -2,7 +2,7 @@ package soccer;
 
 public class Ball {
 	int distance;
-	double degree;
+	int degree;
 	int distanceChange;
 	int degreeChange;
